Day 7 - major intcode changes. scrapped "backward compatibility"
day 7 part 2 requires a machine to stop running and later resume with
same state. therefore it seems pointless to keep the functionality that
makes a copy of the program memory before executing.

also add support for variable length input fifo and output fifo. day 7
does not seem to require more than 2 inputs and 1 output in queue but
it's there now!

Change-Id: Ic41ad46cb91bc7c1ebe99704a6101211dc3cde95